+++ DSN/backend/src/main/java/org/gfa/dsn/utils/SecurityConfig.java
@@ -53,6 +53,7 @@ public class SecurityConfig {
             "/v3/api-docs/**",
             "/v3/api-docs.yaml",
             "/api/v1/auth/**",
+
     };
 
     @Bean
