+++ DSN/frontend/src/app/components/Header.tsx
@@ -1,14 +1,13 @@
+'use client'
 import React from 'react';
-import { Link } from 'react-router-dom';
-
 const Header: React.FC = () => {
     return (
         <header>
-            <a href="#" className="logo">DeveloperHub</a>
+            <div className="logo">DeveloperHub</div>
             <nav className="navigation">
-                <Link to="/key-features">Key Features</Link>
-                <Link to="/developers-say">Developers Say</Link>
-                <Link to="/contact">Contact</Link>
+                <a href="#KeyFeatures">Key Features</a>
+                <a href="#DevelopersSay">Developers Say</a>
+                <a href="#Contact">Contact</a>
             </nav>
         </header>
     );
